Added partial partition table edits
We can now alter an existing partition table.
If the table option on a disk model is set to false then a existing
partition table will be altered instead of writing a complete new
partitiontable.

Resizing of existing partitions in a partitiontable is now also possible
This is set on a per partition basis with the new resize option. If the
resize option is provided than you must also provide a size option
telling us the new size of your partitition.

A new optional argument is provided to the partitiontable. You can now
set the offset attribute on a partition. This offset it the partition
number in the partitiontable. By default we calculate it as the N-th
element of the array of partitions. But you can override that behaviour
by explicitly setting that option to a number